refactor #174:  refactor api fetch method and reduced leaderboard data
