+++ pages/api/timeLeaderboard/index.ts
@@ -3,35 +3,14 @@ import { NextApiRequest, NextApiResponse } from 'next'
 import { formatRoundsData } from '@/utils/formatRoundsData'
 import { findTopFive } from '@/utils/findTopFive'
 import { roundNames } from '@/constants'
-import { mockApiData } from '@/constants/mock'
-import { removeBlacklistedPlayers } from '@/utils/removeBlacklistedPlayers'
+import { fetchData } from '@/utils/fetchData'
 
 export default async function handler(
   req: NextApiRequest,
   res: NextApiResponse,
 ) {
-  const apiKey = process.env.API_KEY
-
   try {
-    if (!apiKey) {
-      throw new Error()
-    }
-
-    let data = []
-
-    if (process.env.NODE_ENV === 'development') {
-      data = mockApiData
-    } else {
-      const response = await fetch(`${apiKey}players`, {
-        method: 'GET',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-      })
-
-      data = await response.json()
-      data = removeBlacklistedPlayers(data)
-    }
+    const data = await fetchData('players')
 
     const formattedData = data.map((elem: ApiPlayerStats) => {
       const saveData = JSON.parse(elem['Save Data'])
